requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg[binary]>=3.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0


# Environment & Configuration
python-dotenv==1.0.0

# Data Validation
pydantic>=2.7.0

# Testing
pytest==7.4.3
httpx==0.25.2


python-multipart